Plataforma B3/requirements.txt
# Dependências principais
streamlit==1.32.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiohttp==3.9.1

# Dependências de desenvolvimento
pytest==7.4.3
black==23.12.1
isort==5.13.2
mypy==1.8.0
pylint==3.0.3 